requirements.txt
gspread==5.12.0
oauth2client==4.1.3
pandas==2.1.4
python-dotenv==1.0.0 